--- bsl/externals/pylsl/version.py
@@ -1,1 +1,0 @@
-__version__ = '1.16.0'
